+++ app/src/main/kotlin/me/xizzhu/android/joshua/core/repository/local/LocalReadingStorage.kt
@@ -18,6 +18,7 @@ package me.xizzhu.android.joshua.core.repository.local
 
 import me.xizzhu.android.joshua.core.Verse
 import me.xizzhu.android.joshua.core.VerseIndex
+import me.xizzhu.android.joshua.core.VerseSearchQuery
 
 interface LocalReadingStorage {
     suspend fun readCurrentVerseIndex(): VerseIndex
@@ -43,5 +44,5 @@ interface LocalReadingStorage {
 
     suspend fun readVerses(translationShortName: String, verseIndexes: List<VerseIndex>): Map<VerseIndex, Verse>
 
-    suspend fun search(translationShortName: String, query: String): List<Verse>
+    suspend fun search(query: VerseSearchQuery): List<Verse>
 }