+++ app/src/main/kotlin/me/xizzhu/android/joshua/reading/toolbar/ReadingToolbar.kt
@@ -32,6 +32,10 @@ class ReadingToolbar : MaterialToolbar {
     constructor(context: Context, attrs: AttributeSet) : super(context, attrs)
     constructor(context: Context, attrs: AttributeSet, defStyleAttr: Int) : super(context, attrs, defStyleAttr)
 
+    init {
+        inflateMenu(R.menu.menu_reading)
+    }
+
     fun initialize(
             requestParallelTranslation: (String) -> Unit,
             removeParallelTranslation: (String) -> Unit,